+++ code/ui/components/dragdrop/Drag.scss
@@ -1,0 +1,3 @@
+Drag {
+
+}
