--- code/ui/components/dragdrop/DragDrop.scss
@@ -1,3 +1,0 @@
-DragDrop {
-
-}
