+++ code/ui/components/dragdrop/Drop.scss
@@ -1,0 +1,3 @@
+Drop {
+
+}
